+++ lib/sym/app/commands/keychain_add_key.rb
@@ -1,5 +1,6 @@
 require 'sym/app/commands/base_command'
 require 'sym/app/keychain'
+require 'sym/errors'
 module Sym
   module App
     module Commands
@@ -11,7 +12,11 @@ module Sym
         try_after :generate_key, :encrypt, :decrypt, :password_protect_key
 
         def execute
+          if Sym.default_key? && Sym.default_key == self.key
+            raise 'Refusing to import key specified in the default key file ' + Sym.default_key_file.italic
+          end
           add_to_keychain_if_needed(self.key)
+          self.key unless opts[:quiet]
         end
       end
     end